requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
python-multipart==0.0.6
python-dotenv==1.0.0
pdfplumber==0.10.3
pandas==2.1.3
openpyxl==3.1.2
requests==2.31.0
aiohttp==3.9.1
beautifulsoup4==4.12.2
langchain==0.1.0
langchain-openai==0.0.2
langchain-community==0.0.10
langgraph==0.0.20
qdrant-client==1.7.0
openai==1.6.1
lxml==4.9.3
pymongo==4.6.0

